feat(eslint): added new `no-extra-functions-in-main` rule
